mailround/requirements.txt
imapclient>=2.1.0
python-dotenv>=0.15.0
msgpack>=1.0.2
jsonschema>=3.2.0